fix: balance error message removed for multisig tx, stxAddress made not required, fixed correct address selection when stxAddress is set
