+++ src/app/common/transactions/stacks/is-sip-10-transfer.ts
@@ -2,7 +2,7 @@ import { ContractCallPayload, StacksTransaction } from '@stacks/transactions';
 
 export function isSip10Transfer(tx: StacksTransaction) {
   if (!tx.payload || !('functionName' in tx.payload)) return false;
-  const payload = tx.payload as ContractCallPayload;
+  const payload = tx.payload;
   return (
     payload.functionName.content === 'transfer' &&
     (payload.functionArgs.length === 3 || payload.functionArgs.length === 4)
