+++ src/app/features/ledger/hooks/use-ledger-latest-route-error.hook.ts
@@ -8,7 +8,7 @@ export function useLatestLedgerError() {
     const state = location.state;
     if (!state || state === null) return null;
     if (typeof state === 'object') {
-      const error = (state as any).latestLedgerError;
+      const error = state.latestLedgerError;
       if (error) return error;
     }
     return null;
