--- src/app/features/stacks-transaction-request/components/stacks-transaction.context.ts
@@ -1,7 +1,0 @@
-import { createContext } from 'react';
-
-import { TransactionPayload } from '@stacks/connect';
-
-export const stacksTransactionContext = createContext<TransactionPayload | null>(null);
-
-export const StacksTransactionProvider = stacksTransactionContext.Provider;
