+++ src/app/features/stacks-transaction-request/hooks/use-transaction-error.ts
@@ -4,6 +4,7 @@ import { ContractCallPayload, TransactionTypes } from '@stacks/connect';
 import BigNumber from 'bignumber.js';
 
 import { useDefaultRequestParams } from '@app/common/hooks/use-default-request-search-params';
+import { initialSearchParams } from '@app/common/initial-search-params';
 import { microStxToStx } from '@app/common/money/unit-conversion';
 import { validateStacksAddress } from '@app/common/stacks-utils';
 import { TransactionErrorReason } from '@app/features/stacks-transaction-request/components/transaction-error/transaction-error';
@@ -11,6 +12,10 @@ import { useCurrentStacksAccountAnchoredBalances } from '@app/query/stacks/balan
 import { useContractInterface } from '@app/query/stacks/contract/contract.hooks';
 import { useCurrentStacksAccount } from '@app/store/accounts/blockchain/stacks/stacks-account.hooks';
 import { useTransactionRequestState } from '@app/store/transactions/requests.hooks';
+
+export function getIsMultisig() {
+  return initialSearchParams.get('isMultisig') === 'true';
+}
 
 export function useTransactionError() {
   const transactionRequest = useTransactionRequestState();
@@ -33,7 +38,7 @@ export function useTransactionError() {
       if ((contractInterface as any)?.isError) return TransactionErrorReason.NoContract;
     }
 
-    if (balances) {
+    if (balances && !getIsMultisig()) {
       const zeroBalance = balances?.stx.unlockedStx.amount.toNumber() === 0;
 
       if (transactionRequest.txType === TransactionTypes.STXTransfer) {
