+++ src/app/pages/rpc-sign-stacks-transaction/rpc-sign-stacks-transaction.tsx
@@ -1,4 +1,3 @@
-import { StacksTransactionProvider } from '@app/features/stacks-transaction-request/components/stacks-transaction.context';
 import { StacksTransactionSigner } from '@app/features/stacks-transaction-request/stacks-transaction-signer';
 import { useRpcSignStacksTransaction } from '@app/pages/rpc-sign-stacks-transaction/use-rpc-sign-stacks-transaction';
 
@@ -6,22 +5,18 @@ export function RpcSignStacksTransaction() {
   const {
     onSignStacksTransaction,
     onCancel,
-    transactionRequest,
     disableFeeSelection,
     stacksTransaction,
     disableNonceSelection,
   } = useRpcSignStacksTransaction();
-  if (!transactionRequest) return null;
 
   return (
-    <StacksTransactionProvider value={transactionRequest}>
-      <StacksTransactionSigner
-        onSignStacksTransaction={onSignStacksTransaction}
-        onCancel={onCancel}
-        stacksTransaction={stacksTransaction}
-        disableFeeSelection={disableFeeSelection}
-        disableNonceSelection={disableNonceSelection}
-      />
-    </StacksTransactionProvider>
+    <StacksTransactionSigner
+      onSignStacksTransaction={onSignStacksTransaction}
+      onCancel={onCancel}
+      stacksTransaction={stacksTransaction}
+      disableFeeSelection={disableFeeSelection}
+      disableNonceSelection={disableNonceSelection}
+    />
   );
 }
