+++ src/app/query/stacks/bns/bns.utils.ts
@@ -41,9 +41,7 @@ async function fetchBnsxName(client: StacksClient, address: string): Promise<str
     });
     if (!res.okay || !res.result) return null;
     const { result } = res;
-    const cv = deserializeCV(result) as OptionalCV<
-      TupleCV<{ name: BufferCV; namespace: BufferCV }>
-    >;
+    const cv = deserializeCV(result);
     if (cv.type === ClarityType.OptionalNone) return null;
     const { name, namespace } = cv.value.data;
     const fullName = `${bytesToAscii(name.buffer)}.${bytesToAscii(namespace.buffer)}`;
@@ -83,7 +81,7 @@ async function fetchBnsxOwner(client: StacksClient, fqn: string): Promise<string
 
   if (!res.okay || !res.result) return null;
   const { result } = res;
-  const cv = deserializeCV(result) as OptionalCV<TupleCV<{ owner: PrincipalCV; id: UIntCV }>>;
+  const cv = deserializeCV(result);
   if (cv.type === ClarityType.OptionalNone) return null;
   const ownerCV = cv.value.data.owner;
   return principalToString(ownerCV);
